tests: move tests to the right package
