+++ plugins/core/src/test/kotlin/de/fayard/refreshVersions/core/TomlLineTest.kt
@@ -1,4 +1,4 @@
-package de.fayard.refreshVersions
+package de.fayard.refreshVersions.core
 
 import de.fayard.refreshVersions.core.internal.TomlLine
 import de.fayard.refreshVersions.core.internal.TomlLine.Kind.Delete
