+++ plugins/core/src/test/kotlin/de/fayard/refreshVersions/core/TomlSectionTest.kt
@@ -1,4 +1,4 @@
-package de.fayard.refreshVersions
+package de.fayard.refreshVersions.core
 
 import de.fayard.refreshVersions.core.internal.Toml
 import io.kotest.core.spec.style.StringSpec
